code/requirements.txt
matplotlib==3.5.2
pandas==1.4.2
python-dotenv==0.20.0
requests==2.26.0
